# kv: re-enable time-bound iterators for RefreshRange request