--- a/pkg/kv/kvserver/batcheval/.gitignore
+++ b/pkg/kv/kvserver/batcheval/.gitignore
@@ -1,0 +1,1 @@
+refresh_range_bench_data_*