--- a/pkg/kv/kvserver/batcheval/cmd_refresh_range_test.go
+++ b/pkg/kv/kvserver/batcheval/cmd_refresh_range_test.go
@@ -220,3 +220,63 @@ func TestRefreshRangeError(t *testing.T) {
 		}
 	})
 }
+
+// TestRefreshRangeTimestampBounds verifies that a RefreshRange treats its
+// RefreshFrom timestamp as exclusive and its txn.ReadTimestamp (i.e. its
+// "RefreshTo" timestamp) as inclusive.
+func TestRefreshRangeTimestampBounds(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	defer log.Scope(t).Close(t)
+
+	ctx := context.Background()
+	db := storage.NewDefaultInMemForTesting()
+	defer db.Close()
+
+	k := roachpb.Key("key")
+	v := roachpb.MakeValueFromString("val")
+	ts1 := hlc.Timestamp{WallTime: 1}
+	ts2 := hlc.Timestamp{WallTime: 2}
+	ts3 := hlc.Timestamp{WallTime: 3}
+
+	// Write to a key at time ts2.
+	require.NoError(t, storage.MVCCPut(ctx, db, nil, k, ts2, v, nil))
+
+	for _, tc := range []struct {
+		from, to hlc.Timestamp
+		expErr   bool
+	}{
+		// Sanity-check.
+		{ts1, ts3, true},
+		// RefreshTo is inclusive, so expect error on collision.
+		{ts1, ts2, true},
+		// RefreshFrom is exclusive, so expect no error on collision.
+		{ts2, ts3, false},
+	} {
+		var resp roachpb.RefreshRangeResponse
+		_, err := RefreshRange(ctx, db, CommandArgs{
+			Args: &roachpb.RefreshRangeRequest{
+				RequestHeader: roachpb.RequestHeader{
+					Key:    k,
+					EndKey: k.Next(),
+				},
+				RefreshFrom: tc.from,
+			},
+			Header: roachpb.Header{
+				Txn: &roachpb.Transaction{
+					TxnMeta: enginepb.TxnMeta{
+						WriteTimestamp: tc.to,
+					},
+					ReadTimestamp: tc.to,
+				},
+				Timestamp: tc.to,
+			},
+		}, &resp)
+
+		if tc.expErr {
+			require.Error(t, err)
+			require.Regexp(t, "encountered recently written committed value", err)
+		} else {
+			require.NoError(t, err)
+		}
+	}
+}