--- a/pkg/roachpb/api.proto
+++ b/pkg/roachpb/api.proto
@@ -1730,8 +1730,9 @@ message RefreshRequest {
   RequestHeader header = 1 [(gogoproto.nullable) = false, (gogoproto.embed) = true];
   reserved 2;
 
-  // refresh_from specifies the lower-bound of the verification. The request verifies that
-  // there's no write in the range [refresh_from, txn.read_timestamp].
+  // refresh_from specifies the exclusive lower-bound of the verification. The
+  // request verifies that there are no writes (committed or provisional) in the
+  // range (refresh_from, txn.read_timestamp].
   util.hlc.Timestamp refresh_from = 3 [(gogoproto.nullable) = false];
 }
 
@@ -1747,8 +1748,9 @@ message RefreshRangeRequest {
   RequestHeader header = 1 [(gogoproto.nullable) = false, (gogoproto.embed) = true];
   reserved 2;
 
-  // refresh_from specifies the lower-bound of the verification. The request verifies that
-  // there's no write in the range [refresh_from, txn.read_timestamp].
+  // refresh_from specifies the exclusive lower-bound of the verification. The
+  // request verifies that there are no writes (committed or provisional) in the
+  // range (refresh_from, txn.read_timestamp].
   util.hlc.Timestamp refresh_from = 3 [(gogoproto.nullable) = false];
 }